removed search bar view and added to game view
